test: seperate test cases
